# packages/control-plane/src/routes/api-keys.ts
import { FastifyInstance } from "fastify";
import { Stratum } from "@stratum-hq/lib";
import { z } from "zod";
import { ForbiddenError } from "@stratum-hq/core";
import { declareTenantScope, fromBodyTenantId, isOperator } from "../middleware/tenant-scope.js";
import { parseWindow } from "../middleware/per-key-rate-limit.js";
import { config } from "../config.js";
import { buildAuditContext } from "./audit-logs.js";
import { legacyValidationErrorBody } from "../middleware/error-handler.js";

const createApiKeySchema = z.object({
  tenant_id: z.string().uuid(),
  name: z.string().optional(),
  rate_limit_max: z.number().int().min(1).max(100_000).optional(),
  rate_limit_window: z.string().regex(/^[1-9]\d*\s+(second|minute|hour|day)s?$/i).optional(),
});

export function createApiKeyRoutes(stratum: Stratum) {
  return async function apiKeyRoutes(app: FastifyInstance): Promise<void> {
    // Scoped keys can only manage keys for their own tenant subtree. On create,
    // the target tenant is in the body; the other routes scope themselves in
    // their handlers.
    declareTenantScope(app, fromBodyTenantId);

    // POST /api/v1/api-keys — Create a new API key (plaintext returned once)
    app.post<{ Body: { tenant_id: string; name?: string } }>(
      "/",
      async (request, reply) => {
        const parsed = createApiKeySchema.safeParse(request.body);
        if (!parsed.success) {
          reply.status(400).send(legacyValidationErrorBody("Invalid request body", parsed.error.issues));
          return;
        }
        const { tenant_id, name, rate_limit_max, rate_limit_window } = parsed.data;
        // A tenant-scoped caller may tighten a key's rate limit but not loosen it
        // beyond the operator's configured default.
        if (!isOperator(request) && (rate_limit_max !== undefined || rate_limit_window !== undefined)) {
          const defaultWindowMs = parseWindow(config.rateLimitWindow);
          const max = rate_limit_max ?? config.rateLimitMax;
          const windowMs = rate_limit_window ? parseWindow(rate_limit_window) : defaultWindowMs;
          if (max > config.rateLimitMax || max * defaultWindowMs > config.rateLimitMax * windowMs) {
            throw new ForbiddenError("Only operator keys may set a per-key rate limit above the configured default");
          }
        }
        const result = await stratum.createApiKey(tenant_id, {
          name,
          rateLimitMax: rate_limit_max,
          rateLimitWindow: rate_limit_window,
        }, undefined, buildAuditContext(request));
        reply.status(201).send(result);
      },
    );

    // GET /api/v1/api-keys — List API keys (optional ?tenant_id=)
    app.get<{ Querystring: { tenant_id?: string } }>("/", async (request, reply) => {
      // Scoped keys can only list keys for their own tenant
      const tenantId = request.apiKey?.tenant_id ?? request.query.tenant_id;
      const keys = await stratum.listApiKeys(tenantId);
      reply.status(200).send(keys);
    });

    // GET /api/v1/api-keys/dormant — List dormant API keys (must be before /:id)
    app.get<{ Querystring: { days?: string } }>("/dormant", async (request, reply) => {
      const raw = request.query.days ? parseInt(request.query.days, 10) : 90;
      const days = Number.isNaN(raw) || raw < 1 ? 90 : Math.min(raw, 365);
      const keys = await stratum.listDormantKeys(days);
      const scopedTenantId = request.apiKey?.tenant_id;
      if (scopedTenantId) {
        const descendants = await stratum.getDescendants(scopedTenantId);
        const allowed = new Set([scopedTenantId, ...descendants.map((d: { id: string }) => d.id)]);
        reply.status(200).send(keys.filter((k: { tenant_id?: string | null }) => k.tenant_id && allowed.has(k.tenant_id)));
        return;
      }
      reply.status(200).send(keys);
    });

    // POST /api/v1/api-keys/:id/rotate — Rotate an API key
    app.post<{ Params: { id: string } }>("/:id/rotate", async (request, reply) => {
      // Verify caller has access to this key's tenant
      if (request.apiKey?.tenant_id) {
        const keys = await stratum.listApiKeys(request.apiKey.tenant_id);
        if (!keys.some(k => k.id === request.params.id)) {
          reply.status(404).send({ error: { code: "NOT_FOUND", message: "API key not found or already revoked" } });
          return;
        }
      }
      const rotateSchema = z.object({ name: z.string().max(255).optional() });
      const parsed = rotateSchema.parse(request.body ?? {});
      const result = await stratum.rotateApiKey(request.params.id, parsed.name, buildAuditContext(request));
      reply.status(201).send(result);
    });

    // DELETE /api/v1/api-keys/:id — Revoke an API key
    app.delete<{ Params: { id: string } }>("/:id", async (request, reply) => {
      // Verify caller has access to this key's tenant
      if (request.apiKey?.tenant_id) {
        const keys = await stratum.listApiKeys(request.apiKey.tenant_id);
        if (!keys.some(k => k.id === request.params.id)) {
          reply.status(404).send({ error: { code: "NOT_FOUND", message: "API key not found or already revoked" } });
          return;
        }
      }
      const revoked = await stratum.revokeApiKey(request.params.id, buildAuditContext(request));
      if (!revoked) {
        reply.status(404).send({ error: { code: "NOT_FOUND", message: "API key not found or already revoked" } });
        return;
      }
      reply.status(204).send();
    });
  };
}