# scripts/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stratum Control Plane API",
    "version": "0.1.0",
    "description": "Universal Tenant Context Engine - Control Plane",
    "contact": {
      "name": "Stratum HQ",
      "url": "https://github.com/stratum-hq"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    { "apiKey": [] },
    { "bearerAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {
                "type": "object",
                "description": "More data about the error, for example the region_id of a REGION_IN_USE error. A 400 VALIDATION_ERROR puts one item per failed field in issues.",
                "properties": {
                  "issues": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/ValidationIssue" }
                  }
                },
                "additionalProperties": true
              },
              "issues": {
                "type": "array",
                "deprecated": true,
                "description": "A copy of details.issues for older clients. It goes away in the next major release.",
                "items": { "$ref": "#/components/schemas/ValidationIssue" }
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
      "ValidationIssue": {
        "type": "object",
        "properties": {
          "path": {
            "type": "array",
            "description": "The keys from the request body root to the field that failed.",
            "items": { "oneOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "message": { "type": "string" },
          "code": { "type": "string", "description": "The zod issue code." }
        },
        "required": ["path", "message", "code"]
      },
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string", "format": "uuid", "nullable": true },
          "name": { "type": "string", "nullable": true },
          "prefix": { "type": "string" },
          "rate_limit_max": { "type": "integer", "nullable": true },
          "rate_limit_window": { "type": "string", "nullable": true },
          "last_used_at": { "type": "string", "format": "date-time", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "revoked_at": { "type": "string", "format": "date-time", "nullable": true }
        },
        "required": ["id", "prefix", "created_at"]
      },
      "ApiKeyCreated": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "key": { "type": "string", "description": "The plaintext API key. Only returned once at creation time." },
          "prefix": { "type": "string" },
          "tenant_id": { "type": "string", "format": "uuid", "nullable": true },
          "name": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "key", "prefix", "created_at"]
      },
      "CreateApiKeyInput": {
        "type": "object",
        "properties": {
          "tenant_id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "rate_limit_max": { "type": "integer", "minimum": 1, "maximum": 100000 },
          "rate_limit_window": { "type": "string", "pattern": "^\\d+\\s+(second|minute|hour|day)s?$" }
        },
        "required": ["tenant_id"]
      },
      "RotateApiKeyInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 255 }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/api-keys": {
      "post": {
        "tags": ["API Keys"],
        "summary": "Create API key",
        "description": "Create a new API key. The plaintext key is only returned once.",
        "operationId": "createApiKey",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateApiKeyInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created API key with plaintext secret",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyCreated" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["API Keys"],
        "summary": "List API keys",
        "description": "List API keys, optionally filtered by tenant.",
        "operationId": "listApiKeys",
        "parameters": [
          { "name": "tenant_id", "in": "query", "schema": { "type": "string", "format": "uuid" }, "description": "Filter by tenant ID" }
        ],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ApiKey" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-keys/dormant": {
      "get": {
        "tags": ["API Keys"],
        "summary": "List dormant API keys",
        "description": "List API keys that have not been used within the specified number of days.",
        "operationId": "listDormantApiKeys",
        "parameters": [
          { "name": "days", "in": "query", "schema": { "type": "integer", "default": 90, "minimum": 1, "maximum": 365 }, "description": "Number of days of inactivity" }
        ],
        "responses": {
          "200": {
            "description": "List of dormant API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ApiKey" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-keys/{id}": {
      "delete": {
        "tags": ["API Keys"],
        "summary": "Revoke API key",
        "description": "Revoke an API key, permanently disabling it.",
        "operationId": "revokeApiKey",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "API key revoked" },
          "404": {
            "description": "API key not found or already revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-keys/{id}/rotate": {
      "post": {
        "tags": ["API Keys"],
        "summary": "Rotate API key",
        "description": "Rotate an API key, revoking the old one and issuing a new one.",
        "operationId": "rotateApiKey",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RotateApiKeyInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New API key with plaintext secret",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyCreated" }
              }
            }
          },
          "404": {
            "description": "API key not found or already revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
...
}